3nd commit

+++ src/main/java/org/example/model/Unit.java
@@ -1,21 +1,23 @@
 package org.example.model;
 
+import java.math.BigDecimal;
+
 public class Unit {
     private String name;
-    private double value;
+    private BigDecimal value;
     public String getName(){
         return this.name;
     }
 
-    public Unit(String name, double value) {
+    public Unit(String name, BigDecimal value) {
         this.name = name;
         this.value = value;
     }
 
-    public double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
-    public void setValue(double value){
+    public void setValue(BigDecimal value){
         this.value = value;
     }
 }
